# Conflicts:
#	README.md
#	android/build.gradle

+++ android/wcdb/src/com/tencent/wcdb/database/SQLiteDatabase.java
@@ -70,18 +70,9 @@ public final class SQLiteDatabase extends SQLiteClosable {
     private static final String TAG = "WCDB.SQLiteDatabase";
 
     static {
-        // To be compatible to frameworks which handle native library loading themselves,
-        // we do a simple test for whether native methods have been registered already.
-        try {
-            SQLiteGlobal.nativeTestJNIRegistration();
-        } catch (UnsatisfiedLinkError e) {
-            // If we reached here, native methods are not registered.
-            System.loadLibrary("wcdb");
-        }
+        // Ensure libmmdb.so is loaded.
+        SQLiteGlobal.loadLib();
     }
-    // Dummy static method to trigger class initialization.
-    // See [JLS 12.4.1](http://docs.oracle.com/javase/specs/jls/se7/html/jls-12.html#jls-12.4.1)
-    public static void loadLib() {}
 
     // Stores reference to all databases opened in the current process.
     // (The referent Object is not used at this time.)
@@ -248,12 +239,6 @@ public final class SQLiteDatabase extends SQLiteClosable {
     public static final int ENABLE_IO_TRACE = 0x00000100;
 
     /**
-     * Open flag: Flag for {@link #openDatabase} that indicates no backup for database files is
-     * done when corruption is detected.
-     */
-    public static final int NO_CORRUPTION_BACKUP = 0x00000200;
-
-    /**
      * Open flag: Flag for {@link #openDatabase} to create the database file if it does not
      * already exist.
      */
@@ -283,7 +268,7 @@ public final class SQLiteDatabase extends SQLiteClosable {
             DatabaseErrorHandler errorHandler) {
         mCursorFactory = cursorFactory;
         mErrorHandler = errorHandler != null ? errorHandler :
-                new DefaultDatabaseErrorHandler((openFlags & NO_CORRUPTION_BACKUP) != 0);
+                new DefaultDatabaseErrorHandler(true);
         mConfigurationLocked = new SQLiteDatabaseConfiguration(path, openFlags);
     }
 
@@ -931,8 +916,7 @@ public final class SQLiteDatabase extends SQLiteClosable {
             throw new IllegalArgumentException("file must not be null");
         }
 
-        boolean deleted = false;
-        deleted |= file.delete();
+        boolean deleted = file.delete();
         deleted |= new File(file.getPath() + "-journal").delete();
         deleted |= new File(file.getPath() + "-shm").delete();
         deleted |= new File(file.getPath() + "-wal").delete();
@@ -946,8 +930,12 @@ public final class SQLiteDatabase extends SQLiteClosable {
                     return candidate.getName().startsWith(prefix);
                 }
             };
-            for (File masterJournal : dir.listFiles(filter)) {
-                deleted |= masterJournal.delete();
+
+            File[] masterJournals = dir.listFiles(filter);
+            if (masterJournals != null) {
+                for (File masterJournal : masterJournals) {
+                    deleted |= masterJournal.delete();
+                }
             }
         }
         return deleted;
@@ -2270,10 +2258,8 @@ public final class SQLiteDatabase extends SQLiteClosable {
             // make sure this database has NO attached databases because sqlite's write-ahead-logging
             // doesn't work for databases with attached databases
             if (mHasAttachedDbsLocked) {
-//                if (Log.isLoggable(TAG, Log.DEBUG)) {
                 Log.i(TAG, "this database: " + mConfigurationLocked.label
                         + " has attached databases. can't  enable WAL.");
-//                }
                 return false;
             }
 
@@ -2351,6 +2337,32 @@ public final class SQLiteDatabase extends SQLiteClosable {
                     throw ex;
                 }
             }
+        }
+    }
+
+    /** Returns the {@link SQLiteChangeListener} object bound to this database.
+     *
+     * @return {@link SQLiteChangeListener} object bound to this database.
+     * @see SQLiteChangeListener
+     * @see #setChangeListener(SQLiteChangeListener, boolean)
+     */
+    public SQLiteChangeListener getChangeListener() {
+        synchronized (mLock) {
+            throwIfNotOpenLocked();
+            return mConnectionPoolLocked.getChangeListener();
+        }
+    }
+
+    /**
+     * Bind a {@link SQLiteChangeListener} object for database change notifications.
+     *
+     * @param listener      listener to be set
+     * @param notifyRowId   whether RowIDs of each modified row should be reported
+     */
+    public void setChangeListener(SQLiteChangeListener listener, boolean notifyRowId) {
+        synchronized (mLock) {
+            throwIfNotOpenLocked();
+            mConnectionPoolLocked.setChangeListener(listener, notifyRowId);
         }
     }
 
@@ -2526,10 +2538,10 @@ public final class SQLiteDatabase extends SQLiteClosable {
                 SQLiteStatement prog = null;
                 try {
                     prog = compileStatement("PRAGMA " + p.first + ".integrity_check(1);");
-                    String rslt = prog.simpleQueryForString();
-                    if (!rslt.equalsIgnoreCase("ok")) {
+                    String result = prog.simpleQueryForString();
+                    if (!DatabaseUtils.objectEquals(result, "ok")) {
                         // integrity_checker failed on main or attached databases
-                        Log.e(TAG, "PRAGMA integrity_check on " + p.second + " returned: " + rslt);
+                        Log.e(TAG, "PRAGMA integrity_check on " + p.second + " returned: " + result);
                         return false;
                     }
                 } finally {
